requirements.txt
arrow==1.2.1
astroid==2.9.0
backports.entry-points-selectable==1.1.1
binaryornot==0.4.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
cookiecutter==1.7.3
datacompy==0.7.3
distlib==0.3.3
filelock==3.4.0
flake8==4.0.1
idna==3.3
isort==5.10.1
Jinja2==3.0.3
jinja2-time==0.2.0
JSON-log-formatter==0.5.0
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.4
ordered-set==4.0.2
pandas==1.3.5
platformdirs==2.4.0
poyo==0.5.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.12.2
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
virtualenv==20.10.0
wrapt==1.13.3
